ka - moved the modal to the CommonsTable for more reasonable testing
